Develop next update

+++ include/staff/templates/list-items.tmpl.php
@@ -6,59 +6,51 @@
         if ($list->getSortMode() == 'SortCol')
             $pageNav->setSlack(1);
         $pageNav->setURL('lists.php?id='.$list->getId().'&a=items');
-        $showing=$pageNav->showing().' '.__('list items');
-        ?>
-    <?php }
-        else $showing = __('Add a few initial items to the list');
+    }
     ?>
     <div style="margin: 5px 0">
-    <div class="pull-left">
-        <input type="search" size="25" id="search" value="<?php
-            echo Format::htmlchars($_POST['search']); ?>"/>
-        <button type="submit" onclick="javascript:
-            event.preventDefault();
-            $.pjax({type: 'POST', data: { search: $('#search').val() }, container: '#pjax-container'});
-            return false;
-"><?php echo __('Search'); ?></button>
-        <?php if ($_POST['search']) { ?>
-        <a href="#" onclick="javascript:
-            $.pjax.reload('#pjax-container'); return false; "
-            ><i class="icon-remove-sign"></i> <?php
-                echo __('clear'); ?></a>
-        <?php } ?>
-    </div>
     <?php if ($list) { ?>
+    <div class="pull-left">
+        <input type="text" placeholder="<?php echo __('Search items'); ?>"
+            data-url="ajax.php/list/<?php echo $list->getId(); ?>/items/search"
+            size="25" id="items-search" value="<?php
+            echo Format::htmlchars($_POST['search']); ?>"/>
+    </div>
     <div class="pull-right">
-        <em style="display:inline-block; padding-bottom: 3px;"><?php echo $showing; ?></em>
-        <?php if ($list->allowAdd()) { ?>
-        <a class="action-button field-config"
+<?php
+if ($list->allowAdd()) { ?>
+        <a class="green button action-button field-config"
             href="#list/<?php
             echo $list->getId(); ?>/item/add">
             <i class="icon-plus-sign"></i>
             <?php echo __('Add New Item'); ?>
         </a>
+<?php
+    if (method_exists($list, 'importCsv')) { ?>
         <a class="action-button field-config"
             href="#list/<?php
             echo $list->getId(); ?>/import">
             <i class="icon-upload"></i>
             <?php echo __('Import Items'); ?>
         </a>
-        <?php } ?>
+<?php
+    }
+} ?>
         <span class="action-button pull-right" data-dropdown="#action-dropdown-more">
             <i class="icon-caret-down pull-right"></i>
             <span ><i class="icon-cog"></i> <?php echo __('More');?></span>
         </span>
         <div id="action-dropdown-more" class="action-dropdown anchor-right">
             <ul>
-                <li><a class="items-action" href="#list/<?php echo $list->getId(); ?>/delete">
-                    <i class="icon-trash icon-fixed-width"></i>
-                    <?php echo __('Delete'); ?></a></li>
                 <li><a class="items-action" href="#list/<?php echo $list->getId(); ?>/disable">
                     <i class="icon-ban-circle icon-fixed-width"></i>
                     <?php echo __('Disable'); ?></a></li>
                 <li><a class="items-action" href="#list/<?php echo $list->getId(); ?>/enable">
                     <i class="icon-ok-sign icon-fixed-width"></i>
                     <?php echo __('Enable'); ?></a></li>
+                <li class="danger"><a class="items-action" href="#list/<?php echo $list->getId(); ?>/delete">
+                    <i class="icon-trash icon-fixed-width"></i>
+                    <?php echo __('Delete'); ?></a></li>
             </ul>
         </div>
     </div>
@@ -69,47 +61,31 @@
 
 
 <?php
-$prop_fields = array();
-if ($list) {
-    foreach ($list->getConfigurationForm()->getFields() as $f) {
-        if (in_array($f->get('type'), array('text', 'datetime', 'phone')))
-            $prop_fields[] = $f;
-        if (strpos($f->get('type'), 'list-') === 0)
-            $prop_fields[] = $f;
-
-        // 4 property columns max
-        if (count($prop_fields) == 4)
-            break;
-    }
-}
+$prop_fields = ($list) ? $list->getSummaryFields() : array();
 ?>
 
     <table class="form_table fixed" width="940" border="0" cellspacing="0" cellpadding="2">
     <thead>
         <tr>
-            <th width="24" nowrap></th>
+            <th width="28" nowrap></th>
             <th><?php echo __('Value'); ?></th>
-<?php foreach ($prop_fields as $F) { ?>
+<?php
+if ($prop_fields) {
+    foreach ($prop_fields as $F) { ?>
             <th><?php echo $F->getLocal('label'); ?></th>
-<?php } ?>
+<?php
+    }
+} ?>
         </tr>
     </thead>
 
-    <tbody <?php if (!isset($_POST['search']) && $list && $list->get('sort_mode') == 'SortCol') { ?>
+    <tbody id="list-items" <?php if (!isset($_POST['search']) && $list && $list->get('sort_mode') == 'SortCol') { ?>
             class="sortable-rows" data-sort="sort-"<?php } ?>>
         <?php
         if ($list) {
             $icon = ($list->get('sort_mode') == 'SortCol')
                 ? '<i class="icon-sort"></i>&nbsp;' : '';
             $items = $list->getAllItems();
-            if ($_POST['search']) {
-                $items->filter(Q::any(array(
-                    'value__contains'=>$_POST['search'],
-                    'extra__contains'=>$_POST['search'],
-                    'properties__contains'=>$_POST['search'],
-                )));
-                $search = true;
-            }
             $items = $pageNav->paginate($items);
             // Emit a marker for the first sort offset ?>
             <input type="hidden" id="sort-offset" value="<?php echo
@@ -125,3 +101,39 @@ if ($list) {
     <div><?php echo __('Page').':'.$pageNav->getPageLinks('items', $pjax_container); ?></div>
 <?php } ?>
 </div>
+<script type="text/javascript">
+$(function() {
+  var last_req;
+  $('input#items-search').typeahead({
+    source: function (typeahead, query) {
+      if (last_req)
+        last_req.abort();
+      var $el = this.$element;
+      var url = $el.data('url')+'?q='+query;
+      last_req = $.ajax({
+        url: url,
+        dataType: 'json',
+        success: function (data) {
+          typeahead.process(data);
+        }
+      });
+    },
+    onselect: function (obj) {
+      var $el = this.$element,
+          url = 'ajax.php/list/{0}/item/{1}/update'
+            .replace('{0}', obj.list_id)
+            .replace('{1}', obj.id);
+      $.dialog(url, [201], function (xhr, resp) {
+        var json = $.parseJSON(resp);
+        if (json && json.success) {
+          if (json.id && json.row) {
+            $('#list-item-' + json.id).replaceWith(json.row);
+          }
+        }
+      });
+      this.$element.val('');
+    },
+    property: "display"
+  });
+});
+</script>
